backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
jsonschema==4.23.0
python-docx==1.1.2
pytest==8.3.4
